+++ src/app/types/account-settings-password.ts
@@ -1,6 +1,9 @@
-export interface AccountPasswordEndpoint {
+import { TwoFactorAuthForm } from './common.endpoint'
+
+export interface AccountPasswordEndpoint extends TwoFactorAuthForm {
   errors: string[]
   password: string
   retypedPassword: string
   oldPassword: string
+  success?: boolean
 }